moved global interfaces to corresponding card components
